Extra restrictie toegevoegd voor rij 11 en 23

+++ Pacman/src/Pacman/Ghost.java
@@ -28,7 +28,8 @@ public class Ghost implements Cloneable {
 		this.type      = type;
 		switch (type) {
 			case GHOST_BLINKY: 
-				position      = new Point2D.Double(13, 11);
+//				position      = new Point2D.Double(13, 11);
+				position      = new Point2D.Double(13, 23);
 				scattertarget = new Point(25, -4);
 				this.direction = PacmanGame.DIR_LEFT;
 			break;
